slicing the code to component
